deps/v8/test/inspector/regress/regress-crbug-1270780-expected.txt

Regression test for crbug.com/1270780.

Running test: testThrowingGenerator
Not pausing for uncaught exception: SUCCESS
Pause no. 1 for caught exception: SUCCESS

Running test: testThrowingNext
Not pausing for uncaught exception: SUCCESS
Pause no. 1 for caught exception: SUCCESS

Running test: testThrowingReturn
Not pausing for uncaught exception: SUCCESS
Pause no. 1 for caught exception: SUCCESS

Running test: testThrowingLoopBody
Not pausing for uncaught exception: SUCCESS
Pause no. 1 for caught exception: SUCCESS

Running test: testThrowingLoopBodyAndThrowingReturn
Not pausing for uncaught exception: SUCCESS
Pause no. 1 for caught exception: SUCCESS
Pause no. 2 for caught exception: SUCCESS
